+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project follows semantic versioning.
 
 ### Unreleased
+- [fixed] References to generic types.
 
 ### 1.0.0 (2023-02-20)
 - [fixed] Bug when repeating a type in an unnamed variant.